[#1174] Move model classes out of the CompactBlockProcessor
* [#1174] SbS: Move model classes out of the CompactBlockProcessor

* Move SyncingResult out of the processor
