+++ sdk-lib/src/test/java/cash/z/ecc/android/sdk/fixture/ScanRangeFixture.kt
@@ -6,7 +6,8 @@ import cash.z.ecc.android.sdk.model.BlockHeight
 import cash.z.ecc.android.sdk.model.ZcashNetwork
 
 object ScanRangeFixture {
-    internal val DEFAULT_CLOSED_RANGE = ZcashNetwork.Testnet.saplingActivationHeight..ZcashNetwork.Testnet.saplingActivationHeight + 9
+    internal val DEFAULT_CLOSED_RANGE =
+        ZcashNetwork.Testnet.saplingActivationHeight..ZcashNetwork.Testnet.saplingActivationHeight + 9
     internal val DEFAULT_PRIORITY = SuggestScanRangePriority.Verify.priority
 
     internal fun new(
